Updated to Cats Effect 3.7, Scala Native 0.5, and dotty-cps-async 1.0. Changed name
