+++ project/plugins.sbt
@@ -1,5 +1,5 @@
 addSbtPlugin("org.typelevel" % "sbt-typelevel" % "0.8.0")
 addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.19.0")
-addSbtPlugin("org.scala-native" % "sbt-scala-native" % "0.4.17")
+addSbtPlugin("org.scala-native" % "sbt-scala-native" % "0.5.8")
 addSbtPlugin("org.portable-scala" % "sbt-scala-native-crossproject" % "1.3.2")
 addSbtPlugin("de.heikoseeberger" % "sbt-header" % "5.10.0")
